fix(lock): guard stale-reap path against live in-proc holders on other threads
A cloned AppContext sharing the same Arc<in_proc> could fall through to
try_reap_stale_lock even while a different thread held an active LockGuard.
If the lock file appeared stale (age > LOCK_STALE_AFTER or missing host
field), the reaper would delete it and the second thread would overwrite
the ref-count entry, breaking mutual exclusion while the first guard was
still live.

Fix: before calling try_reap_stale_lock, check in_proc under the mutex.
Any entry for the requested name means this process has a live holder on a
different thread — return LOCK_BUSY immediately without touching the file.

Also extract lock internals (LockMetadata, is_lock_stale, current_hostname,
pid_status and their unit tests) into src/state/lock.rs to keep mod.rs
under the 800-line file-size limit.

Add cloned_context_different_thread_not_reaped_as_stale test to exercise
the new guard path directly.
